Moved the new e2e test to different location
