+++ test/e2e/app_namespaces_test.go
@@ -4,10 +4,8 @@ import (
 	"context"
 	"testing"
 
-	"github.com/argoproj/gitops-engine/pkg/health"
 	. "github.com/argoproj/gitops-engine/pkg/sync/common"
 	"github.com/stretchr/testify/assert"
-	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
@@ -104,52 +102,4 @@ func TestDeletingNamespacedAppStuckInSync(t *testing.T) {
 		Then().
 		// delete is successful
 		Expect(DoesNotExist())
-}
-
-// Given application is set with --sync-option CreateNamespace=true and --sync-option ServerSideApply=true
-//
-//		application --dest-namespace exists
-//
-//	Then, --dest-namespace is created with server side apply
-//		  	application is synced and healthy with resource
-//		  	application resources created with server side apply in the newly created namespace.
-func TestNamespacedNamespaceAutoCreationWithServerSideApply(t *testing.T) {
-	SkipOnEnv(t, "OPENSHIFT")
-	namespace := "guestbook-ui-with-server-side-apply"
-	defer func() {
-		if !t.Skipped() {
-			_, err := Run("", "kubectl", "delete", "namespace", namespace)
-			assert.NoError(t, err)
-		}
-	}()
-
-	ctx := Given(t)
-	ctx.
-		SetAppNamespace(AppNamespace()).
-		Timeout(30).
-		Path("guestbook").
-		When().
-		CreateFromFile(func(app *Application) {
-			app.Spec.SyncPolicy = &SyncPolicy{
-				SyncOptions: SyncOptions{"CreateNamespace=true", "ServerSideApply=true"},
-			}
-		}).
-		Then().
-		Expect(NoNamespace(namespace)).
-		When().
-		AppSet("--dest-namespace", namespace).
-		Sync().
-		Then().
-		Expect(Success("")).
-		Expect(Namespace(namespace, func(app *Application, ns *v1.Namespace) {
-			assert.NotContains(t, ns.Annotations, "kubectl.kubernetes.io/last-applied-configuration")
-		})).
-		Expect(SyncStatusIs(SyncStatusCodeSynced)).
-		Expect(HealthIs(health.HealthStatusHealthy)).
-		Expect(OperationPhaseIs(OperationSucceeded)).Expect(ResourceHealthWithNamespaceIs("Deployment", "guestbook-ui", namespace, health.HealthStatusHealthy)).
-		Expect(ResourceHealthWithNamespaceIs("Deployment", "guestbook-ui", namespace, health.HealthStatusHealthy)).
-		Expect(ResourceSyncStatusWithNamespaceIs("Deployment", "guestbook-ui", namespace, SyncStatusCodeSynced)).
-		Expect(OperationPhaseIs(OperationSucceeded)).Expect(ResourceHealthWithNamespaceIs("Service", "guestbook-ui", namespace, health.HealthStatusHealthy)).
-		Expect(ResourceHealthWithNamespaceIs("Service", "guestbook-ui", namespace, health.HealthStatusHealthy)).
-		Expect(ResourceSyncStatusWithNamespaceIs("Service", "guestbook-ui", namespace, SyncStatusCodeSynced))
 }
